pyproject.toml
[project]
name = "voinux"
version = "1.0.0"
description = "Privacy-focused, offline-capable voice typing solution for Linux with GPU-accelerated speech recognition"
readme = "README.md"
authors = [
    {name = "Voinux Contributors"}
]
license = {text = "MIT"}
requires-python = ">=3.12"
keywords = ["voice", "speech-recognition", "transcription", "whisper", "linux", "accessibility"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

dependencies = [
    "faster-whisper>=1.0.0",
    "torch>=2.0.0",
    "soundcard>=0.4.2",
    "pyaudio>=0.2.13",
    "webrtcvad>=2.0.10",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
cuda = [
    "nvidia-cublas-cu12",
    "nvidia-cudnn-cu12",
]
rocm = [
    # ROCm PyTorch must be installed separately from pytorch.org
    # See: https://pytorch.org/get-started/locally/ for ROCm installation
]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "pre-commit>=3.0.0",
]

[project.scripts]
voinux = "voinux.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --cov=voinux --cov-report=term-missing"

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.black]
line-length = 100
target-version = ["py312"]
